Refactor: extract DistWorkerManager from DistScheduler
Separate worker pool management from the DAG scheduling engine:

- New DistWorkerManager class owns WorkerThread instances, handles
  add/start/stop/pick_idle/any_busy — extracted from DistScheduler
- DistScheduler becomes a pure DAG engine: ready→dispatch→completion→fanout.
  It calls manager->pick_n_idle() and worker->dispatch() but does not own
  or create WorkerThread instances
- DistWorker owns the WorkerManager as a member, wires the completion
  callback (manager → scheduler.worker_done), starts manager before
  scheduler, stops in reverse order

This is a structural preparation for ForkProxy and inline/fork mode
support per the hierarchical runtime refactor plan.
