+++ python/bindings/CMakeLists.txt
@@ -25,6 +25,7 @@ set(DIST_SOURCES
     ${DIST_SRC}/dist_orchestrator.cpp
     ${DIST_SRC}/dist_sub_worker.cpp
     ${DIST_SRC}/dist_chip_process.cpp
+    ${DIST_SRC}/dist_worker_manager.cpp
     ${DIST_SRC}/dist_scheduler.cpp
     ${DIST_SRC}/dist_worker.cpp
 )
